requirements.txt
fastapi
uvicorn
jinja2
confluent-kafka
python-dotenv
